+++ src/main/java/data_access/AbstractDataAccessObject.java
@@ -2,8 +2,8 @@ package data_access;
 
 import java.util.ArrayList;
 
-import entity.CommonUser;
-import entity.User;
+import entity.player.CommonUser;
+import entity.player.User;
 
 /**
  * An abstact DAO. This is used for any future DAO these are
