+++ src/main/java/data_access/nonpersisting/InMemoryUserDataAccessObject.java
@@ -1,0 +1,190 @@
+package data_access.nonpersisting;
+
+import java.io.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import entity.player.CommonUser;
+import entity.player.User;
+import entity.player.UserFactory;
+import use_case.change_password.ChangePasswordUserDataAccessInterface;
+import use_case.leaderboard.LeaderboardUserDataAccessInterface;
+import use_case.login.LoginUserDataAccessInterface;
+import use_case.logout.LogoutUserDataAccessInterface;
+import use_case.signup.SignupUserDataAccessInterface;
+
+/**
+ * In-memory implementation of the DAO for storing user data. This implementation does
+ * NOT persist data between runs of the program.
+ */
+public class InMemoryUserDataAccessObject implements SignupUserDataAccessInterface,
+        LoginUserDataAccessInterface,
+        ChangePasswordUserDataAccessInterface,
+        LogoutUserDataAccessInterface, LeaderboardUserDataAccessInterface {
+
+    private final Map<String, User> users = new HashMap<>();
+
+    private String currentUsername;
+
+    @Override
+    public boolean existsByName(String identifier) {
+        return users.containsKey(identifier);
+    }
+
+    @Override
+    public void save(User user) {
+        users.put(user.getName(), user);
+    }
+
+    @Override
+    public User get(String username) {
+        return users.get(username);
+    }
+
+    @Override
+    public void changePassword(User user) {
+        // Replace the old entry with the new password
+        users.put(user.getName(), user);
+    }
+
+    @Override
+    public User getCurrentUser() {
+        return null;
+    }
+
+    @Override
+    public void setCurrentUsername(String name) {
+        this.currentUsername = name;
+    }
+
+    @Override
+    public void setCurrentUser(User user) {
+        this.currentUsername = user.getName();
+    }
+
+    @Override
+    public ArrayList<CommonUser> returnAllUsers() {
+        final ArrayList<CommonUser> allUsers = new ArrayList<>();
+        for (User user : users.values()) {
+            allUsers.add(new CommonUser(user.getName(), user.getPassword(), user.getPoints(), user.getNumberOfGames()));
+        }
+        return allUsers;
+    }
+
+    @Override
+    public String getCurrentUsername() {
+        return this.currentUsername;
+    }
+
+    /**
+     * DAO for user data implemented using a File to persist the data.
+     */
+    public static class FileUserDataAccessObject implements SignupUserDataAccessInterface,
+                                                     LoginUserDataAccessInterface,
+                                                     ChangePasswordUserDataAccessInterface {
+
+        private static final String HEADER = "username,password";
+
+        private final File csvFile;
+        private final Map<String, Integer> headers = new LinkedHashMap<>();
+        private final Map<String, User> accounts = new HashMap<>();
+        private String currentUsername;
+
+        public FileUserDataAccessObject(String csvPath, UserFactory userFactory) throws IOException {
+
+            csvFile = new File(csvPath);
+            headers.put("username", 0);
+            headers.put("password", 1);
+
+            if (csvFile.length() == 0) {
+                save();
+            }
+            else {
+
+                try (BufferedReader reader = new BufferedReader(new FileReader(csvFile))) {
+                    final String header = reader.readLine();
+
+                    if (!header.equals(HEADER)) {
+                        throw new RuntimeException(String.format("header should be%n: %s%but was:%n%s", HEADER, header));
+                    }
+
+                    String row;
+                    while ((row = reader.readLine()) != null) {
+                        final String[] col = row.split(",");
+                        final String username = String.valueOf(col[headers.get("username")]);
+                        final String password = String.valueOf(col[headers.get("password")]);
+                        final User user = userFactory.create(username, password);
+                        accounts.put(username, user);
+                    }
+                }
+            }
+        }
+
+        private void save() {
+            final BufferedWriter writer;
+            try {
+                writer = new BufferedWriter(new FileWriter(csvFile));
+                writer.write(String.join(",", headers.keySet()));
+                writer.newLine();
+
+                for (User user : accounts.values()) {
+                    final String line = String.format("%s,%s",
+                            user.getName(), user.getPassword());
+                    writer.write(line);
+                    writer.newLine();
+                }
+
+                writer.close();
+
+            }
+            catch (IOException ex) {
+                throw new RuntimeException(ex);
+            }
+        }
+
+        @Override
+        public void save(User user) {
+            accounts.put(user.getName(), user);
+            this.save();
+        }
+
+        @Override
+        public User get(String username) {
+            return accounts.get(username);
+        }
+
+        @Override
+        public void setCurrentUsername(String name) {
+            this.currentUsername = name;
+        }
+
+        @Override
+        public void setCurrentUser(User user) {
+
+        }
+
+        @Override
+        public String getCurrentUsername() {
+            return this.currentUsername;
+        }
+
+        @Override
+        public boolean existsByName(String identifier) {
+            return accounts.containsKey(identifier);
+        }
+
+        @Override
+        public void changePassword(User user) {
+            // Replace the User object in the map
+            accounts.put(user.getName(), user);
+            save();
+        }
+
+        @Override
+        public User getCurrentUser() {
+            return null;
+        }
+    }
+}
