+++ src/main/java/data_access/streetview/StreetViewDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.streetview;
 
 /**
  * Data Access Object for the StreetView Use Case.
