+++ src/main/java/entity/map/Map2D.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.map;
 
 import java.io.IOException;
 import java.nio.file.Files;
