+++ src/main/java/entity/player/UserFactory.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.player;
 
 /**
  * Factory for creating users.
