+++ src/main/java/entity/points/PointsCalculator.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.points;
 
 /**
  * Calculates the points for the game.
