+++ src/main/java/use_case/change_password/ChangePasswordInteractor.java
@@ -1,7 +1,7 @@
 package use_case.change_password;
 
-import entity.User;
-import entity.UserFactory;
+import entity.player.User;
+import entity.player.UserFactory;
 
 /**
  * The Change Password Interactor.
