+++ src/main/java/use_case/change_password/ChangePasswordUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.change_password;
 
-import entity.User;
+import entity.player.User;
 
 /**
  * The interface of the DAO for the Change Password Use Case.
