+++ src/main/java/use_case/login/LoginInteractor.java
@@ -1,6 +1,6 @@
 package use_case.login;
 
-import entity.User;
+import entity.player.User;
 
 /**
  * The Login Interactor.
