+++ src/main/java/use_case/login/LoginUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.login;
 
-import entity.User;
+import entity.player.User;
 
 /**
  * DAO for the Login Use Case.
