+++ src/main/java/use_case/signup/SignupUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.signup;
 
-import entity.User;
+import entity.player.User;
 
 /**
  * DAO for the Signup Use Case.
