+++ src/main/java/view/game/Map2DView.java
@@ -1,4 +1,4 @@
-package view;
+package view.game;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
