+++ src/main/java/view/game/MapView.java
@@ -1,4 +1,4 @@
-package view;
+package view.game;
 
 import interface_adapter.streetview_map.StreetViewMapController;
