+++ src/main/java/view/game/PointsCalculatorView.java
@@ -1,4 +1,4 @@
-package view;
+package view.game;
 
 import java.awt.*;
 import java.awt.event.ActionEvent;
