+++ src/main/java/view/game/RoundView.java
@@ -1,4 +1,4 @@
-package view;
+package view.game;
 
 import java.awt.Color;
 import java.awt.Component;
