+++ src/test/java/use_case/login/LoginInteractorTest.java
@@ -1,12 +1,10 @@
 package use_case.login;
 
-import data_access.InMemoryUserDataAccessObject;
-import entity.CommonUserFactory;
-import entity.User;
-import entity.UserFactory;
+import data_access.nonpersisting.InMemoryUserDataAccessObject;
+import entity.player.CommonUserFactory;
+import entity.player.User;
+import entity.player.UserFactory;
 import org.junit.jupiter.api.Test;
-
-import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.*;
