+++ src/test/java/use_case/logout/LogoutInteractorTest.java
@@ -1,9 +1,9 @@
 package use_case.logout;
 
-import data_access.InMemoryUserDataAccessObject;
-import entity.CommonUserFactory;
-import entity.User;
-import entity.UserFactory;
+import data_access.nonpersisting.InMemoryUserDataAccessObject;
+import entity.player.CommonUserFactory;
+import entity.player.User;
+import entity.player.UserFactory;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
